+++ tidy3d/components/geometry/base.py
@@ -26,6 +26,7 @@ from ...packaging import check_import, verify_packages_import
 from ..autograd import AutogradFieldMap, TracedCoordinate, TracedSize, get_static
 from ..autograd.derivative_utils import DerivativeInfo, integrate_within_bounds
 from ..base import Tidy3dBaseModel, cached_property
+from ..data.dataset import ElectromagneticFieldDataset
 from ..transformation import RotationAroundAxis
 from ..types import (
     ArrayFloat2D,
@@ -1430,6 +1431,71 @@ class Geometry(Tidy3dBaseModel, ABC):
     def compute_derivatives(self, derivative_info: DerivativeInfo) -> AutogradFieldMap:
         """Compute the adjoint derivatives for this object."""
         raise NotImplementedError(f"Can't compute derivative for 'Geometry': '{type(self)}'.")
+
+    def der_at_centers(
+        self,
+        der_map: ElectromagneticFieldDataset,
+        centers: np.ndarray,  # (N, 3)
+    ) -> xr.Dataset:
+        """Compute the value of an ``ElectromagneticFieldDataset`` at a set of edge centers."""
+
+        xs, ys, zs = centers.T
+        edge_index_dim = "edge_index"
+
+        interp_kwargs = {}
+        for dim, centers_dim in zip("xyz", centers.T):
+            # only include dims where the data has more than 1 coord, to avoid warnings and errors
+            coords_data = der_map[f"E{dim}"].coords
+            if np.array(coords_data).size > 1:
+                interp_kwargs[dim] = xr.DataArray(centers_dim, dims=edge_index_dim)
+
+        components = {}
+        for fld_name, arr in der_map.items():
+            components[fld_name] = arr.interp(**interp_kwargs).sum("f")
+
+        return xr.Dataset(components)
+
+    def project_in_basis(
+        self,
+        der_dataset: xr.Dataset,
+        basis_vector: np.ndarray,
+    ) -> xr.DataArray:
+        """Project a derivative dataset along a supplied basis vector."""
+
+        value = 0.0
+        for coeffs, dim in zip(basis_vector.T, "xyz"):
+            value += coeffs * der_dataset[f"E{dim}"]
+        return value
+
+    def unpop_axis_vect(self, ax_coords: np.ndarray, plane_coords: np.ndarray) -> np.ndarray:
+        """Combine coordinate along axis with coordinates on the plane tangent to the axis.
+
+        ax_coords.shape == [N] or [N, 1]
+        plane_coords.shape == [N, 2]
+        return shape == [N, 3]
+
+        """
+        arr_xyz = self.unpop_axis(ax_coords, plane_coords.T, axis=self.axis)
+        arr_xyz = np.stack(arr_xyz, axis=-1)
+        return arr_xyz
+
+    def pop_axis_vect(self, coord: np.ndarray) -> Tuple[np.ndarray, Tuple[np.ndarray, np.ndarray]]:
+        """Combine coordinate along axis with coordinates on the plane tangent to the axis.
+
+        coord.shape == [N, 3]
+        return shape == ([N], [N, 2]
+
+        """
+
+        arr_axis, arrs_plane = self.pop_axis(coord.T, axis=self.axis)
+        arrs_plane = np.array(arrs_plane).T
+
+        return arr_axis, arrs_plane
+
+    @staticmethod
+    def normalize_vect(arr: np.ndarray) -> np.ndarray:
+        """normalize an array shaped (N, d) along the `d` axis and return (N, 1)."""
+        return arr / np.linalg.norm(arr, axis=-1)[..., None]
 
     def _as_union(self) -> List[Geometry]:
         """Return a list of geometries that, united, make up the given geometry."""