+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for differentiation with respect to `GeometryGroup.geometries` elements.
+- Support for differentiation with respect to `Cylinder.radius` and `Cylinder.center`.
 - Users can now export `SimulationData` to MATLAB `.mat` files with the `to_mat_file` method.
 
 ### Changed